:bug: 杀虫啦

+++ src/main/java/cn/tycoding/admin/controller/ArticleController.java
@@ -8,6 +8,7 @@ import cn.tycoding.admin.enums.ResultEnums;
 import cn.tycoding.admin.service.ArticleService;
 import cn.tycoding.admin.service.ArticleTagsService;
 import cn.tycoding.admin.service.CategoryService;
+import cn.tycoding.admin.utils.CheckValue;
 import com.alibaba.fastjson.JSON;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.ui.Model;
@@ -48,21 +49,20 @@ public class ArticleController {
     public Result findByPage(Article article,
                              @RequestParam(value = "pageCode", required = false) Integer pageCode,
                              @RequestParam(value = "pageSize", required = false) Integer pageSize) {
-        if (pageCode != null && pageSize != null) {
+        if (CheckValue.checkPage(pageCode, pageSize)) {
             return new Result(StatusCode.SUCCESS, articleService.findByPage(article, pageCode, pageSize));
+        } else {
+            return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
         }
-        return new Result(StatusCode.ERROR, ResultEnums.ERROR);
     }
 
     @RequestMapping(value = "/findById", method = RequestMethod.GET)
     public Result findById(@RequestParam("id") Long id, Model model) {
-        if (id == null || id == 0) {
-            return new Result(StatusCode.ERROR, ResultEnums.ERROR);
-        } else {
+        if (CheckValue.checkId(id)) {
             Article article = articleService.findById(id);
-            if (article != null) {
+            if (CheckValue.checkObj(article)) {
                 if (article.getId() != 0) {
-                    List<String> tags = new ArrayList<String>();
+                    List<String> tags = new ArrayList<>();
                     List<Tags> tagsList = articleTagsService.findByArticleId(article.getId());
                     for (Tags t : tagsList) {
                         tags.add(t.gettName());
@@ -74,11 +74,12 @@ public class ArticleController {
             }
             return new Result(StatusCode.SUCCESS, article);
         }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 
     @RequestMapping(value = "/findTags", method = RequestMethod.GET)
     public Result findTags(@RequestParam("id") Long id) {
-        if (id != null) {
+        if (CheckValue.checkId(id)) {
             List<String> list = new ArrayList<String>();
             List<Tags> tagsList = articleTagsService.findByArticleId(id);
             for (Tags t : tagsList) {
@@ -86,7 +87,7 @@ public class ArticleController {
             }
             return new Result(StatusCode.SUCCESS, list);
         } else {
-            return new Result(StatusCode.ERROR, ResultEnums.ERROR);
+            return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
         }
     }
 
@@ -97,34 +98,43 @@ public class ArticleController {
 
     @RequestMapping(value = "/save", method = RequestMethod.POST)
     public Result save(@RequestBody Article article) {
-        try {
-            articleService.save(article);
-            return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return new Result(StatusCode.ERROR, ResultEnums.ERROR);
+        if (CheckValue.checkObj(article)) {
+            try {
+                articleService.save(article);
+                return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new Result(StatusCode.ERROR, e.getMessage());
+            }
         }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 
     @RequestMapping(value = "/update", method = RequestMethod.PUT)
     public Result update(@RequestBody Article article) {
-        try {
-            articleService.update(article);
-            return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return new Result(StatusCode.ERROR, ResultEnums.ERROR);
+        if (CheckValue.checkObj(article)) {
+            try {
+                articleService.update(article);
+                return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new Result(StatusCode.ERROR, e.getMessage());
+            }
         }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 
     @RequestMapping(value = "/delete", method = RequestMethod.POST)
     public Result delete(@RequestBody Long... ids) {
-        try {
-            articleService.delete(ids);
-            return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return new Result(StatusCode.ERROR, ResultEnums.ERROR);
+        if (CheckValue.checkIds(ids)) {
+            try {
+                articleService.delete(ids);
+                return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new Result(StatusCode.ERROR, e.getMessage());
+            }
         }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 }
