+++ src/main/java/cn/tycoding/admin/controller/CommentsController.java
@@ -5,6 +5,7 @@ import cn.tycoding.admin.dto.StatusCode;
 import cn.tycoding.admin.entity.Comments;
 import cn.tycoding.admin.enums.ResultEnums;
 import cn.tycoding.admin.service.CommentsService;
+import cn.tycoding.admin.utils.CheckValue;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -34,10 +35,10 @@ public class CommentsController {
     public Result findByPage(Comments comments,
                              @RequestParam(value = "pageCode", required = false) Integer pageCode,
                              @RequestParam(value = "pageSize", required = false) Integer pageSize) {
-        if (pageCode != null && pageSize != null) {
+        if (CheckValue.checkPage(pageCode, pageSize)) {
             return new Result(StatusCode.SUCCESS, commentsService.findByPage(comments, pageCode, pageSize));
         }
-        return new Result(StatusCode.ERROR, ResultEnums.ERROR);
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 
     /**
@@ -48,10 +49,10 @@ public class CommentsController {
      */
     @RequestMapping(value = "/findCountByArticleId", method = RequestMethod.GET)
     public Result findCountByArticleId(@RequestParam("articleId") Long articleId) {
-        if (articleId != null && articleId != 0) {
+        if (CheckValue.checkId(articleId)) {
             return new Result(StatusCode.SUCCESS, commentsService.findCountByArticle(articleId));
         }
-        return new Result(StatusCode.ERROR, ResultEnums.ERROR);
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 
     /**
@@ -60,59 +61,66 @@ public class CommentsController {
      * @param pageCode  当前页
      * @param pageSize  每页显示的记录数
      * @param articleId 当前访问的文章ID
-     * @param sort 分类，规定：sort=0表示文章详情页的评论信息；sort=1表示友链页的评论信息；sort=2表示关于我页的评论信息
+     * @param sort      分类，规定：sort=0表示文章详情页的评论信息；sort=1表示友链页的评论信息；sort=2表示关于我页的评论信息
      * @return
      */
     @RequestMapping(value = "/findCommentsList", method = RequestMethod.GET)
     public Result findCommentsList(@RequestParam(value = "pageCode", required = false) Integer pageCode,
                                    @RequestParam(value = "pageSize", required = false) Integer pageSize,
                                    @RequestParam(value = "articleId", required = false) Integer articleId) {
-        if (pageCode != null && pageSize != null && articleId != null && articleId != 0) {
+        if (CheckValue.checkPageIds(pageCode, pageSize, articleId)) {
             return new Result(StatusCode.SUCCESS, commentsService.findCommentsList(pageCode, pageSize, articleId, 0));
-        } else {
-            return new Result(StatusCode.ERROR, ResultEnums.ERROR);
         }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 
     @RequestMapping(value = "/findById", method = RequestMethod.GET)
     public Result findById(@RequestParam("id") Long id) {
         if (id != null && id != 0) {
             return new Result(StatusCode.SUCCESS, commentsService.findById(id));
-        } else {
-            return new Result(StatusCode.ERROR, ResultEnums.ERROR);
         }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 
     @RequestMapping(value = "/save", method = RequestMethod.POST)
     public Result save(@RequestBody Comments comments) {
-        try {
-            commentsService.save(comments);
-            return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return new Result(StatusCode.ERROR, ResultEnums.ERROR);
+        if (CheckValue.checkObj(comments)) {
+            try {
+                commentsService.save(comments);
+                return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new Result(StatusCode.ERROR, e.getMessage());
+            }
         }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 
     @RequestMapping(value = "/update", method = RequestMethod.PUT)
     public Result update(@RequestBody Comments comments) {
-        try {
-            commentsService.update(comments);
-            return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return new Result(StatusCode.ERROR, ResultEnums.ERROR);
+        if (CheckValue.checkObj(comments)) {
+            try {
+                commentsService.update(comments);
+                return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new Result(StatusCode.ERROR, e.getMessage());
+            }
         }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 
     @RequestMapping(value = "/delete", method = RequestMethod.POST)
     public Result delete(@RequestBody Long... ids) {
-        try {
-            commentsService.delete(ids);
-            return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return new Result(StatusCode.ERROR, ResultEnums.ERROR);
+        if (CheckValue.checkIds(ids)) {
+            try {
+                commentsService.delete(ids);
+                return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new Result(StatusCode.ERROR, e.getMessage());
+            }
         }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 }
