+++ src/main/java/cn/tycoding/admin/controller/LinksController.java
@@ -5,6 +5,7 @@ import cn.tycoding.admin.dto.StatusCode;
 import cn.tycoding.admin.entity.Links;
 import cn.tycoding.admin.enums.ResultEnums;
 import cn.tycoding.admin.service.LinksService;
+import cn.tycoding.admin.utils.CheckValue;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -34,44 +35,59 @@ public class LinksController {
     public Result findByPage(Links links,
                              @RequestParam(value = "pageCode", required = false) Integer pageCode,
                              @RequestParam(value = "pageSize", required = false) Integer pageSize) {
-        return new Result(StatusCode.SUCCESS, linksService.findByPage(links, pageCode, pageSize));
+        if (CheckValue.checkPage(pageCode, pageSize)) {
+            return new Result(StatusCode.SUCCESS, linksService.findByPage(links, pageCode, pageSize));
+        }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 
     @RequestMapping(value = "/findById", method = RequestMethod.GET)
     public Result findById(@RequestParam("id") Long id) {
-        return new Result(StatusCode.SUCCESS, linksService.findById(id));
+        if (CheckValue.checkId(id)) {
+            return new Result(StatusCode.SUCCESS, linksService.findById(id));
+        }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 
     @RequestMapping(value = "/save", method = RequestMethod.POST)
     public Result save(@RequestBody Links links) {
-        try {
-            linksService.save(links);
-            return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return new Result(StatusCode.ERROR, ResultEnums.ERROR);
+        if (CheckValue.checkObj(links)) {
+            try {
+                linksService.save(links);
+                return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new Result(StatusCode.ERROR, e.getMessage());
+            }
         }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 
     @RequestMapping(value = "/update", method = RequestMethod.PUT)
     public Result update(@RequestBody Links links) {
-        try {
-            linksService.update(links);
-            return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return new Result(StatusCode.ERROR, ResultEnums.ERROR);
+        if (CheckValue.checkObj(links)) {
+            try {
+                linksService.update(links);
+                return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new Result(StatusCode.ERROR, e.getMessage());
+            }
         }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 
     @RequestMapping(value = "/delete", method = RequestMethod.POST)
     public Result delete(@RequestBody Long... ids) {
-        try {
-            linksService.delete(ids);
-            return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return new Result(StatusCode.ERROR, ResultEnums.ERROR);
+        if (CheckValue.checkIds(ids)) {
+            try {
+                linksService.delete(ids);
+                return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new Result(StatusCode.ERROR, e.getMessage());
+            }
         }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 }
