+++ src/main/java/cn/tycoding/admin/controller/TagsController.java
@@ -5,6 +5,7 @@ import cn.tycoding.admin.dto.StatusCode;
 import cn.tycoding.admin.entity.Tags;
 import cn.tycoding.admin.enums.ResultEnums;
 import cn.tycoding.admin.service.TagsService;
+import cn.tycoding.admin.utils.CheckValue;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -21,7 +22,7 @@ public class TagsController {
     private TagsService tagsService;
 
     @RequestMapping(value = "/findAllCount", method = RequestMethod.GET)
-    public Result findAllCount(){
+    public Result findAllCount() {
         return new Result(StatusCode.SUCCESS, tagsService.findAllCount());
     }
 
@@ -34,44 +35,59 @@ public class TagsController {
     public Result findByPage(Tags tags,
                              @RequestParam(value = "pageCode", required = false) Integer pageCode,
                              @RequestParam(value = "pageSize", required = false) Integer pageSize) {
-        return new Result(StatusCode.SUCCESS, tagsService.findByPage(tags, pageCode, pageSize));
+        if (CheckValue.checkPage(pageCode, pageSize)) {
+            return new Result(StatusCode.SUCCESS, tagsService.findByPage(tags, pageCode, pageSize));
+        }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 
     @RequestMapping(value = "/findById", method = RequestMethod.GET)
     public Result findById(@RequestParam("id") Long id) {
-        return new Result(StatusCode.SUCCESS, tagsService.findById(id));
+        if (CheckValue.checkId(id)) {
+            return new Result(StatusCode.SUCCESS, tagsService.findById(id));
+        }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 
     @RequestMapping(value = "/save", method = RequestMethod.POST)
     public Result save(@RequestBody Tags tags) {
-        try {
-            tagsService.save(tags);
-            return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return new Result(StatusCode.ERROR, ResultEnums.ERROR);
+        if (CheckValue.checkObj(tags)) {
+            try {
+                tagsService.save(tags);
+                return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new Result(StatusCode.ERROR, e.getMessage());
+            }
         }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 
     @RequestMapping(value = "/update", method = RequestMethod.PUT)
     public Result update(@RequestBody Tags tags) {
-        try {
-            tagsService.update(tags);
-            return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return new Result(StatusCode.ERROR, ResultEnums.ERROR);
+        if (CheckValue.checkObj(tags)) {
+            try {
+                tagsService.update(tags);
+                return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new Result(StatusCode.ERROR, e.getMessage());
+            }
         }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 
     @RequestMapping(value = "/delete", method = RequestMethod.POST)
     public Result delete(@RequestBody Long... ids) {
-        try {
-            tagsService.delete(ids);
-            return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return new Result(StatusCode.ERROR, ResultEnums.ERROR);
+        if (CheckValue.checkIds(ids)) {
+            try {
+                tagsService.delete(ids);
+                return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new Result(StatusCode.ERROR, e.getMessage());
+            }
         }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 }
