+++ src/main/java/cn/tycoding/admin/controller/UserController.java
@@ -6,6 +6,7 @@ import cn.tycoding.admin.dto.StatusCode;
 import cn.tycoding.admin.entity.User;
 import cn.tycoding.admin.enums.ResultEnums;
 import cn.tycoding.admin.service.UserService;
+import cn.tycoding.admin.utils.CheckValue;
 import org.apache.shiro.SecurityUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -33,29 +34,38 @@ public class UserController {
     public Result findByPage(User user,
                              @RequestParam(value = "pageCode", required = false) Integer pageCode,
                              @RequestParam(value = "pageSize", required = false) Integer pageSize) {
-        return new Result(StatusCode.SUCCESS, userService.findByPage(user, pageCode, pageSize));
+        if (CheckValue.checkPage(pageCode, pageSize)) {
+            return new Result(StatusCode.SUCCESS, userService.findByPage(user, pageCode, pageSize));
+        }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 
     @RequestMapping(value = "/findById", method = RequestMethod.GET)
     public Result findById(@RequestParam("id") Long id) {
-        return new Result(StatusCode.SUCCESS, userService.findById(id));
+        if (CheckValue.checkId(id)) {
+            return new Result(StatusCode.SUCCESS, userService.findById(id));
+        }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 
     @RequestMapping(value = "/save", method = RequestMethod.POST)
     public Result save(@RequestBody User user) {
-        try {
-            userService.save(user);
-            return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return new Result(StatusCode.ERROR, ResultEnums.ERROR);
+        if (CheckValue.checkObj(user)) {
+            try {
+                userService.save(user);
+                return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new Result(StatusCode.ERROR, e.getMessage());
+            }
         }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 
     @RequestMapping("/update")
     public Result update(@RequestBody User user) {
-        try {
-            if (user != null) {
+        if (CheckValue.checkObj(user)) {
+            try {
                 if (user.getPassword() != null && !"".equals(user.getPassword()) && user.getCheckPass() != null && !"".equals(user.getCheckPass())) {
                     //说明是更新密码操作
                     User u = userService.findByName((String) SecurityUtils.getSubject().getPrincipal());
@@ -74,27 +84,28 @@ public class UserController {
                         // check success
                         userService.update(user);
                     }
-                } else {
-                    userService.update(user);
                 }
+                userService.update(user);
                 return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
-            } else {
-                return new Result(StatusCode.ERROR, ResultEnums.ERROR);
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new Result(StatusCode.ERROR, e.getMessage());
             }
-        } catch (Exception e) {
-            e.printStackTrace();
-            return new Result(StatusCode.ERROR, ResultEnums.ERROR);
         }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 
     @RequestMapping(value = "/delete", method = RequestMethod.POST)
     public Result delete(@RequestBody Long... ids) {
-        try {
-            userService.delete(ids);
-            return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return new Result(StatusCode.ERROR, ResultEnums.ERROR);
+        if (CheckValue.checkIds(ids)) {
+            try {
+                userService.delete(ids);
+                return new Result(StatusCode.SUCCESS, ResultEnums.SUCCESS);
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new Result(StatusCode.ERROR, ResultEnums.ERROR);
+            }
         }
+        return new Result(StatusCode.PARAMETER_ERROR, ResultEnums.PARAMETER_ERROR);
     }
 }
